# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DC",
        "description": "Endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "description": "Obtiene una lista paginada de todos los productos.",
                "operationId": "getAllProducts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de productos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "pages": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error inesperado al obtener la lista de productos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto con los datos proporcionados.",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "brand_id",
                                    "category_id",
                                    "measurement_unit_id",
                                    "name",
                                    "sale_price",
                                    "purchase_price",
                                    "minimum_stock"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "measurement_unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "measurement_unit_name": {
                                        "type": "string",
                                        "example": "Unidad"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop Gamer"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "barcode_path": {
                                        "type": "string",
                                        "example": "/images/barcodes/1234567890123.png"
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1499.99
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.5
                                    },
                                    "minimum_stock": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "brand_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El ID de la marca es obligatorio."
                                                    }
                                                },
                                                "category_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El ID de la categoría es obligatorio."
                                                    }
                                                },
                                                "measurement_unit_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El ID de la unidad de medida es obligatorio."
                                                    }
                                                },
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El nombre del producto es obligatorio."
                                                    }
                                                },
                                                "sale_price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El precio de venta es obligatorio."
                                                    }
                                                },
                                                "purchase_price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El precio de compra es obligatorio."
                                                    }
                                                },
                                                "minimum_stock": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El stock mínimo es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error inesperado al crear el producto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "title": "Product Resource",
                "description": "Estructura del recurso de un producto",
                "required": [
                    "id",
                    "brand_id",
                    "category_id",
                    "measurement_unit_id",
                    "name",
                    "sale_price",
                    "purchase_price",
                    "minimum_stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101,
                        "nullable": false
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": false
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": false
                    },
                    "measurement_unit_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": false
                    },
                    "measurement_unit_name": {
                        "type": "string",
                        "example": "Unidad",
                        "nullable": false
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop Gamer",
                        "nullable": false
                    },
                    "barcode": {
                        "type": "string",
                        "example": "1234567890123",
                        "nullable": true
                    },
                    "barcode_path": {
                        "type": "string",
                        "example": "/images/barcodes/1234567890123.png",
                        "nullable": true
                    },
                    "sale_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1499.99,
                        "nullable": false
                    },
                    "purchase_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1200.5,
                        "nullable": false
                    },
                    "minimum_stock": {
                        "type": "integer",
                        "example": 5,
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}